Implement resized event and refactor event loop
SDL does not guarantee that WaitEvent blocks consistently. For instance,
currently if any JoySticks (SDL name for controllers) are detected, and
the joystick/controller system is enabled, SDL will fallback to polling
for events and waiting 1ms in case no event is available before polling
again. So I might as well try something different with the architecture.

I actually thought about just dumping this idea as it was honestly kind
of dumb, but I really wan't to see how far I can stretch it. It is a
fun side-project anyway.